🚚 Split config_types.go
This splits the config type declarations into one for the operator and
one for the platform.
